File Upload is completed...

+++ Node_Practice/src/module-24/controllers/feed.controller.js
@@ -1,18 +1,24 @@
 const { validationResult } = require('express-validator');
 const Post = require('../models/post.model');
+const User = require('../models/user.model');
+const fs = require('fs');
+const path = require('path');
 
 exports.getFeeds = (req, res, next) => {
   let totalItems = 0;
   const page = req.query.page;
   const perPage = req.query.perPage || 2;
+  let creator;
   Post.countDocuments()
     .then((total) => {
       totalItems = total;
       return Post.find()
+        .populate('creator', 'name email')
         .skip((page - 1) * perPage)
         .limit(perPage);
     })
     .then((result) => {
+      // result.creator.name = 'Mr Deba';
       res.status(200).json({
         result: 'success message',
         posts: result,
@@ -29,8 +35,8 @@ exports.getFeeds = (req, res, next) => {
 
 exports.postFeeds = (req, res, next) => {
   const error = validationResult(req);
-  console.table(req);
   const file = req.file;
+  let postRecord;
 
   if (!file) {
     const error = new Error('Sorry no file selected');
@@ -47,19 +53,27 @@ exports.postFeeds = (req, res, next) => {
   const post = {
     title: req.body.title,
     content: req.body.content,
-    creator: JSON.parse(req.body.creator),
+    creator: req.userId,
     imageUrl: req.file.filename,
   };
 
-  console.table(post);
+  console.log(post);
   // DATABASE operation
-  return Post(post)
+  Post(post)
     .save()
     .then((result) => {
-      console.log('Saved', res);
+      postRecord = result;
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.post.push(postRecord._id);
+      return user.save();
+    })
+    .then(() => {
       return res.status(201).json({
         result: 'success message',
-        receivedPost: result,
+        receivedPost: post,
+        creator: { id: req.userId },
       });
     })
     .catch((error) => {
@@ -92,18 +106,84 @@ exports.getSinglePost = (req, res, next) => {
     });
 };
 
-exports.deleteSinglePost = (req, res, next) => {
+exports.deleteSinglePost = async (req, res, next) => {
   const postId = req.params.id;
-  Post.findByIdAndDelete(postId)
+
+  Post.findById(postId)
     .then((post) => {
       if (!post) {
         const error = new Error('No post is there with this id');
         error.setStatus = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId.toString()) {
+        const error = new Error('You can not delete this post');
+        error.setStatus = 422;
+        throw error;
+      }
+      return Post.findByIdAndDelete(postId);
+    })
+    .then((result) => {
+      return User.findById(result.creator);
+    })
+    .then((user) => {
+      user.post.pull(postId);
+      return user.save();
+    })
+    .then((userDetail) => {
+      res.status(200).json({
+        result: 'successfully deleted message',
+        userDetail,
+      });
+    })
+    .catch((err) => {
+      if (!err.status && !err.setStatus) {
+        err.setStatus = 500;
+      }
+      next(err);
+    });
+};
+
+exports.updateSinglePost = (req, res, next) => {
+  const postId = req.params.id;
+  const newFileSelected = req.file;
+
+  Post.findById(postId)
+    .then((post) => {
+      if (!post) {
+        const error = new Error('No post is there with this id');
+        error.setStatus = 404;
+        throw error;
+      }
+      if (post.creator.toString() !== req.userId.toString()) {
+        const error = new Error('You can not edit this post');
+        error.setStatus = 422;
+        throw error;
+      }
+      if (newFileSelected) {
+        console.log(path.join(__dirname, 'images', post.imageUrl));
+        fs.unlink(
+          path.join(__dirname, '..', 'images', post.imageUrl),
+          (err) => {
+            if (!err) {
+              console.log('file removed');
+            }
+          }
+        );
+      }
+      return Post.findByIdAndUpdate(postId, {
+        $set: {
+          content: req.body.content,
+          title: req.body.title,
+          imageUrl: newFileSelected
+            ? newFileSelected.filename
+            : req.body.imageUrl,
+        },
+      });
+    })
+    .then(() => {
       res.status(200).json({
         result: 'success message',
-        post,
       });
     })
     .catch((err) => {
